bin/speciation/requirements.txt
pandas
PyVCF
numpy